+++ arx/inc/cur_def.h
@@ -74,6 +74,7 @@ WINDOW *cur_init (int around, int *ysz, int *xsz, unsigned char *col);
 void cur_exit (int status);
 int cur_getk ();
 int cur_getp (WINDOW *w);
+void cur_puts(int y, int x, char *s, int w);
 int cur_gets (WINDOW *w, int y, int x, int width, int att, char *s, int pos, char *legal, int maxlength, int *chg, char *(*f4edit)(char *fml, char *pos));
 int cur_satt (WINDOW *w, unsigned int attrib);
 int cur_syst (WINDOW *w, char *cmd, int prompt);
