+++ arx/src/lib_fsiz.c
