+++ arx/src/str_tlen.c
@@ -1,4 +1,6 @@
 /* str_tlen.c
+ * UTF-8 to ISO-8859-1/ISO-8859-15 mapper.
+ * Return 0..255 for valid ISO-8859-15 code points, 256 otherwise.
  * length without trailing whitespace
  * substring with printf %*.*s capabilities from utf-8 mbstowcs
  * cc -I../inc -DTEST str_tlen.c -olll
@@ -8,10 +10,143 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <wchar.h>
 #include <arx_def.h>
 #include <str_def.h>
+#include <cur_def.h>
+#include "oem2iso.h"
+
+char *arx_vers(char *rcs) {
+#ifdef UTF8
+static char version[132];
+snprintf(version, 130, "%-.72s w=%c", rcs, cur_utf8 ? '1' : '0');
+return version;
+#else
+return rcs;
+#endif
+}
+
+void oem858_to_iso885915(char *output)
+{
+  while (*output) {
+    *output = oem858[(unsigned char)(*output)];
+    output++;
+  }
+}
+
+void iso885915_to_oem858(char *output)
+{
+  while (*output) {
+    *output = iso885915[(unsigned char)(*output)];
+    output++;
+  }
+}
 
 int cur_utf8 = 0;
+
+/* Convert an UTF-8 string to UCODE
+ * All invalid sequences are ignored.
+ * Note: output == input is allowed,
+ * but   input < output < input + length is not.
+ * Output has to have room for (length+1) chars, including the trailing NUL byte.
+ */
+int utf8_to_ucode(wchar_t *output, char *input, int length) {
+    wchar_t                   *out = output;
+    const unsigned char       *in  = (const unsigned char *)input;
+    const unsigned char *const end = (const unsigned char *)input + strlen(input);
+    unsigned int               c;
+
+    while (in < end)
+        if (*in < 128)
+            *(out++) = *(in++); /* Valid codepoint */
+        else
+        if (*in < 192)
+            break;
+/*          in++;                * 10000000 .. 10111111 are invalid */
+        else
+        if (*in < 224) {        /* 110xxxxx 10xxxxxx */
+            if (in + 1 >= end)
+                break;
+            if ((in[1] & 192U) == 128U) {
+                c =   (((unsigned int)(in[0] & 0x1FU)) << 6U)
+                    | ((unsigned int)(in[1] & 0x3FU)) ;
+                *(out++) = c;
+            }
+            in += 2;
+
+        } else
+        if (*in < 240) {        /* 1110xxxx 10xxxxxx 10xxxxxx */
+            if (in + 2 >= end)
+                break;
+            if ((in[1] & 192U) == 128U &&
+                (in[2] & 192U) == 128U) {
+                c =   (((unsigned int)(in[0] & 0x0FU)) << 12U)
+                    | (((unsigned int)(in[1] & 0x3FU)) << 6U)
+                    |  ((unsigned int)(in[2] & 0x3FU)) ;
+                *(out++) = c;
+            }
+            in += 3;
+
+        } else
+        if (*in < 248) {        /* 11110xxx 10xxxxxx 10xxxxxx 10xxxxxx */
+            if (in + 3 >= end)
+                break;
+            if ((in[1] & 192U) == 128U &&
+                (in[2] & 192U) == 128U &&
+                (in[3] & 192U) == 128U) {
+                c =   (((unsigned int)(in[0] & 0x07U)) << 18U)
+                    | (((unsigned int)(in[1] & 0x3FU)) << 12U)
+                    | (((unsigned int)(in[2] & 0x3FU)) << 6U)
+                    |  ((unsigned int)(in[3] & 0x3FU)) ;
+                *(out++) = c;
+            }
+            in += 4;
+
+        } else
+        if (*in < 252) {        /* 111110xx 10xxxxxx 10xxxxxx 10xxxxxx 10xxxxxx */
+            if (in + 4 >= end)
+                break;
+            if ((in[1] & 192U) == 128U &&
+                (in[2] & 192U) == 128U &&
+                (in[3] & 192U) == 128U &&
+                (in[4] & 192U) == 128U) {
+                c =   (((unsigned int)(in[0] & 0x03U)) << 24U)
+                    | (((unsigned int)(in[1] & 0x3FU)) << 18U)
+                    | (((unsigned int)(in[2] & 0x3FU)) << 12U)
+                    | (((unsigned int)(in[3] & 0x3FU)) << 6U)
+                    |  ((unsigned int)(in[4] & 0x3FU)) ;
+                *(out++) = c;
+            }
+            in += 5;
+
+        } else
+        if (*in < 254) {        /* 1111110x 10xxxxxx 10xxxxxx 10xxxxxx 10xxxxxx 10xxxxxx */
+            if (in + 5 >= end)
+                break;
+            if ((in[1] & 192U) == 128U &&
+                (in[2] & 192U) == 128U &&
+                (in[3] & 192U) == 128U &&
+                (in[4] & 192U) == 128U &&
+                (in[5] & 192U) == 128U) {
+                c =   (((unsigned int)(in[0] & 0x01U)) << 30U)
+                    | (((unsigned int)(in[1] & 0x3FU)) << 24U)
+                    | (((unsigned int)(in[2] & 0x3FU)) << 18U)
+                    | (((unsigned int)(in[3] & 0x3FU)) << 12U)
+                    | (((unsigned int)(in[4] & 0x3FU)) << 6U)
+                    |  ((unsigned int)(in[5] & 0x3FU)) ;
+                *(out++) = c;
+            }
+            in += 6;
+
+        } else
+            in++;               /* 11111110 and 11111111 are invalid */
+
+    /* Terminate the output string. */
+    *out = 0;
+
+    if (in < end) return -1;
+    else          return out - output;
+}
 
 static int utf162utf8(char* out, int c) {
 int bits;
@@ -66,7 +201,6 @@ for (p = s; *p; p++) {
   if (z > 0 && n >= f+z) break;
   n++;
  }
-// fprintf(stderr,"%d %c %4x %4x %d\n",(!cur_utf8),*p,*p & 0xC0,0x80,n);
 }
 n -= f;
 if (z > 0) {
