+++ mc/src/mcellstr.c
@@ -1,6 +1,6 @@
 /* mcellstr.c */
 
-//#pragma GCC diagnostic ignored "-Wformat-overflow"
+/* #pragma GCC diagnostic ignored "-Wformat-overflow" */
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -72,12 +72,8 @@ static void textstring (char *instring, char *outstring,
 			int col, unsigned int format, int formatting)
 /* Sets the string representation of text */
 {
-#ifdef NOUTF8
-char		*ljust = "%-*s", *rjust = "%*s ";
-#else
 char		*ljust = "%-s", *rjust = "%s ";
 char *adjusted;
-#endif
 char		*just	= ljust;
 int width	= colwidth[col];
 
@@ -86,28 +82,18 @@ switch (*instring)
  {
  case STRRIGHT:
   if (formatting) {just = rjust; width--;}
-#ifndef NOUTF8
   if (formatting) width = -width;
-#endif
 	/*FALLTHRU*/
  case STRLEFT:
  case STRCENTER:
 	instring++;
  default:
-#ifdef NOUTF8
-	sprintf(outstring, just, width, instring);
-#else
 	adjusted = str_sub(NULL, instring, 0, width, (formatting==FORMAT || formatting==FPRINT) ? colwidth[col] : 0);
-//  fprintf(stderr, ":%s: = str_sub(NULL, :%s:, 0, %d, 0)\n", adjusted, instring, width);
 	sprintf(outstring, just, adjusted);
   free(adjusted);
-#endif
 	switch (formatting) {
 	 case FORMAT:
 	 case FPRINT:
-#ifdef NOUTF8
-		outstring[colwidth[col]] = '\0';
-#endif
 		break;
 	 }
 	break;
@@ -288,10 +274,7 @@ int		pos;
 CELLPTR		cp;
 unsigned char	typ = EMPTY;
 int		visible = FALSE;
-#undef NOUTF8
-#ifndef NOUTF8
 char *adjusted;
-#endif
 
 s = s1+1;
 cp = cell (col, row);
@@ -321,19 +304,10 @@ if (cp == NULL)
 		p = celltext(newcol, row);
 		for (pos = 0; newcol < col; pos += colwidth[newcol++])
       ; /* left side truncate */
-#ifdef NOUTF8
-		p += min(strlen(p), (size_t)pos);
-		visible = (*p != '\0');
-		strncpy(temp, p, colwidth[col]);
-		temp[colwidth[col]] = '\0';
-		sprintf(s, "%s%*s", temp, (int)(colwidth[col] - strlen(temp)), "");
-#else
     adjusted = str_sub(NULL, p, pos, colwidth[col], colwidth[col]);
     visible = !strcspn(adjusted, " ");
-// fprintf(stderr, "%d:%s: = str_sub(NULL, :%s:, %d, %d, %d)\n", visible, adjusted, p, pos, colwidth[col], colwidth[col]);
     strcpy(s, adjusted);
     free(adjusted);
-#endif
 		*color = TEXTCOLOR;
 		}
 	}
