+++ mc/src/mcput.c
@@ -24,11 +24,6 @@ va_list	args;
 int	oldx, oldy;
 char	s[MAXINPUT+1];
 int color;
-#ifndef NOUTF8
-extern int cur_utf8;
-char *out = NULL;
-wchar_t t[MAXINPUT+1];
-#endif
 
 color = mcode2att(colcode & TYPEM);
 switch (colcode & BIMASK) {
@@ -44,17 +39,7 @@ va_end (args);
 cur_satt (stdscr, color);
 setcolor(colcode);
 s[SCREENWIDTH+1] = '\0';
-
-#ifndef NOUTF8
-if (cur_utf8) {
-  out = str_sub(out, s, 0, width, 0);
-  utf8_to_ucode(t, out, MAXINPUT);
-  move (y, x);
-  addwstr(t);
-  free(out);
-} else
-#endif
-mvprintw (y, x, "%-*s", width, s);
+cur_puts(y, x, s, width);
 
 #ifdef DEBUG
 fprintf (stderr, "put: y:%2d x:%2d s:\"%-*s\">%08x\n", y, x, width, s, color);
